fix indent bug

+++ app/controllers/IndentController.php
@@ -46,6 +46,9 @@ class IndentController extends BaseController {
 		if($indent->user_id != $user->id)
 			return Response::json(array('errCode' => 2,'message' => '你没有此操作权限!'));
 
+		if($indent->status > 0)
+			return Response::json(array('errCode' => 3,'message' => '你已付款不能再取消订单!'));
+
 		$content = '用户' . $user->name . '取消了你的' . $indent->product->name . '商品的订单!';
 		$receiver = $indent->product->shop->user_id;
 		if($indent->delete())
@@ -60,7 +63,7 @@ class IndentController extends BaseController {
 			return Response::json(array('errCode' => 0,'message' => '取消订单成功!'));
 		}	
 
-		return Response::json(array('errCode' => 3,'message' => '取消订单失败!'));
+		return Response::json(array('errCode' => 4,'message' => '取消订单失败!'));
 	}
 
 	public function getIndent()
